enable type check during eval time
